Scaffolds test scheme
In order to avoid having all tests in the same target
(which would allow compile-time enumeration), we use
the instantiation of runtime object across multiple
targets to allow the enumeration of tests.

This adds an O(N) overhead when the test is launched,
since the test table is built before the main() is
called. Hopefully, the effect should be negligible
under multiple hundreds of thousands of tests.