Update introspection/introspection.h, bytewise/bytewise.h
Updates introspection to make $tag compatible with types
containing a comma (as in the case of templates). Those types
can now be parenthesized in the macro, hence treated as a
single argument. The parentheses are removed using the new trait
`argument` in `introspection`.

Expands $bytewise macro to account for base classes.