Update network/tcp
Set `SO_REUSEADDR` to true by default, so that port can be reused
once the previous connection isn't active anymore.

Add `set` and `get` for `reuse` (see above) and `domain`.
`get <domain>` returns whether the socket is IPv4 or IPv6.

Fix `bind` so that, if only the port is specified, it uses an
address of the correct domain (IPv4 or IPv6).