Update network/connection
Refactor `secure` into `secure <peer>`, `secure <client>` and
`secure <server>`. Allowing to secure a connection in all the
following situations: when both hosts know the public key of the
other, when only one hosr knows the public key of the other and
when both hosts don't know the public key of the other.