Update async/aggregators
Adds full implementation of `any` aggregator. To be tested
more deeply.

Includes a patch to handle an optimizer-related
issue with copy-captures in lambdas: the optimizer apparently
feels free to optimize out a copy-capture, even if the variable
is used after a `co_await` (which could, in principle, result in
the variable being destroyed).